Split Interact.cs into two scripts
Because of the way it was set up, every GameObject that had a the interact script was checking if the player camera hit any other interactable object. I moved that check to a script on the player object, because this check only need to be done once per frame.
