+++ sg1/Assets/Scenes/Prefabs/Player/Player.prefab
@@ -18,6 +18,7 @@ GameObject:
   - component: {fileID: 590713874040520505}
   - component: {fileID: 825168620957131300}
   - component: {fileID: 7939238632856961546}
+  - component: {fileID: 9119347396111180124}
   m_Layer: 0
   m_Name: Player
   m_TagString: Player
@@ -274,3 +275,17 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   doorController: {fileID: 0}
+--- !u!114 &9119347396111180124
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 3570673063139267787}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 0c184df789c9c444991ec35802c3a718, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  interactionDistance: 3
+  playerCamera: {fileID: 0}
