--- sg1/Assets/Scripts/Interact.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: c69f940a079771b47934b516e6830665
