+++ sg1/Assets/Scripts/Interactable.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: b511c45b3bf016f49aabd4a5011ae3f6
