+++ sg1/Assets/Scripts/Interacter.cs
@@ -1,12 +1,14 @@
 using UnityEngine;
 
-public class Interactable : MonoBehaviour
+public class Interacter : MonoBehaviour
 {
-    public InventoryItem item; // Reference to the item to be added to the inventory
-
     public float interactionDistance = 3f; // Distance within which you can interact
     public Camera playerCamera; // Reference to the player's camera
 
+    private void Start()
+    {
+        playerCamera = GetComponent<Camera>();
+    }
     void Update()
     {
         if (Input.GetKeyDown(KeyCode.F)) // Check if the "F" key is pressed
@@ -23,21 +25,4 @@ public class Interactable : MonoBehaviour
         }
     }
 
-    public void Interact()
-    {
-        InventoryManagement inventoryManager = FindAnyObjectByType<InventoryManagement>();
-        if (inventoryManager != null)
-        {
-            if (gameObject.activeSelf == true)
-            {
-                inventoryManager.AddItemToHotBar(item); // Add the item to the inventory
-            }
-            gameObject.SetActive(false);
-            Destroy(gameObject); // Remove the item from the scene
-        }
-        else
-        {
-            Debug.LogWarning("InventoryManager not found!");
-        }
-    }
 }
