+++ sg1/Assets/Scripts/Interacter.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 0c184df789c9c444991ec35802c3a718
